git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13568 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/createmesh.cpp
@@ -712,7 +712,7 @@ for (i = FACES.nFaces, faceP = FACES.faces.Buffer (), nFace = 0; i; i--, faceP++
 	}
 //insert each face's vertices' indices in the vertex index buffer
 memset (bTags, 0xFF, gameData.segs.nVertices * sizeof (bTags [0]));
-if (gameData.segs.nFaceVerts > FACES.faceVerts.Length ())
+if ((uint32_t) gameData.segs.nFaceVerts > FACES.faceVerts.Length ())
 	FACES.faceVerts.Resize (gameData.segs.nFaceVerts);
 gameData.segs.nFaceVerts = 0;
 for (i = FACES.nFaces, faceP = FACES.faces.Buffer (), nFace = 0; i; i--, faceP++, nFace++) {
